refactor: 💡 Cleaned unnecesary comments
